# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SC Advogados",
        "description": "Visão dos endpoints para o projeto SC Advogados",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/contract-block": {
            "get": {
                "tags": [
                    "contract-block"
                ],
                "summary": "Lista todas bloco de contrato",
                "description": "Lista todas bloco de contrato",
                "operationId": "213c0bd190140a690f620aa37215ab7c",
                "responses": {
                    "200": {
                        "description": "Lista todas bloco de contrato"
                    }
                }
            },
            "put": {
                "tags": [
                    "contract-block"
                ],
                "summary": "Edita um bloco de contrato",
                "operationId": "a4f8646d2786bf5bba45815a40eb582e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "question": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "condition_type": {
                                        "type": "integer"
                                    },
                                    "field_id": {
                                        "type": "integer"
                                    },
                                    "field_value": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "integer"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "contract-block"
                ],
                "summary": "Cadastra um novo bloco de contrato",
                "operationId": "6884a976dee0b751b078cf5d9ab05316",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "question": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "condition_type": {
                                        "type": "integer"
                                    },
                                    "field_id": {
                                        "type": "integer"
                                    },
                                    "field_value": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "integer"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/contract-block/{id}": {
            "get": {
                "tags": [
                    "contract-block"
                ],
                "summary": "Exibe um bloco de contrato",
                "operationId": "dd674a2a324c71c3d51b7b0482cb1b70",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "contract-block"
                ],
                "summary": "Remove um bloco de contrato",
                "operationId": "8c520fadaafbd33771bd3cc4dc1906ae",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/contract-type": {
            "get": {
                "tags": [
                    "contract-type"
                ],
                "summary": "Lista todos os tipos de contratos",
                "description": "Lista todos os tipos de contratos",
                "operationId": "aa6bbb4688878d845f4b7dd8c2d236a1",
                "responses": {
                    "200": {
                        "description": "Lista todos os tipos de contratos cadastrados"
                    }
                }
            },
            "put": {
                "tags": [
                    "contract-type"
                ],
                "summary": "Editado um tipo de contrato",
                "operationId": "f31c844cf7b21a337478dc8c7e992172",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "contract-type"
                ],
                "summary": "Cadastra um novo tipo de contrato",
                "operationId": "b9be84612a106dea217115da82beca32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/contract-type/{id}": {
            "get": {
                "tags": [
                    "contract-type"
                ],
                "summary": "Exibe um tipo de contrato",
                "operationId": "2d1e0226935ec8779c1ac51f3b985862",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "contract-type"
                ],
                "summary": "Remove um tipo de contrato",
                "operationId": "66e8ff7809e7240e63990ae7069c41b0",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}